Reorganize, variable name API call, safe metadata
